Added check to prevent loading code unnecessarily
Since these metaboxes are only meant to be used on the post edit
screen, I think it’s best to avoid loading any of this code in other
parts of wp-admin.

This check will allow the metaboxes to display only on the post.php and
post-new.php screens. Normally I prefer get_current_screen() for this
but it looks like we’re hooked in too early to use it.