Added wp_current_filter check
This should prevent the unformatted shortcode data from being displayed
in excerpts. I think in general this function could use some more
checking to account for other weird use cases, but this does seem to do
the trick for now.

For a good example of additional checks that should probably be added,
take a look at this function in Jetpack:
https://github.com/Automattic/jetpack/blob/5fc81afe593f0ee1fa14c4a1f4844
0e5ab11a59f/modules/sharedaddy/sharing-service.php#L491-L636

Fixes #34