Some more tweaks to the caching.
A couple thoughts on caching:

The wp_cache_add calls should be wp_cache_set. The difference between
the two is that if the key is already set, wp_cache_add will not update
the value. The add will just fail. wp_cache_set will always update the
value.
Each time you set the cache, set an expiration time. Yes, we'd ideally
cache it forever then invalidate it only when needed; however, I've
found it to be best practice to always set an expiration time because
naturally updating your cache can fix issues that would otherwise be
big bugs.
Let's not use URLs for cache keys. There could potentially be illegal
chars in the string or the key could be too long. Just hash it.