Report Filesize on Podcast Download
The Content-Length header was not being included on the podcast download, since the $file is a URL.  I don't know if this is the best way to handle it, but it works for me.  I wondered if there was a way to use $this->get_file_size(...) but that looked like it downloads the file from the URL to find its size - not too efficient?  I just used the saved value from the saved meta information.